+++ marketplaces/amazon_parser.py
@@ -359,7 +359,7 @@ class AmazonParser:
         else:
             return 'Unknown'
 
-    def get_smaller_size(self) -> int | str:
+    def get_smaller_size(self) -> float | str:
         """Получение меньшего значения в размере товара для сортировки"""
         files = self.__search_link_to_file()
         try:
@@ -368,10 +368,10 @@ class AmazonParser:
                     if file['name'] != 'File Not Found':
                         size_from_name = file['name'].split(" ")[0]
                         size = size_from_name.split("x")
-                        width = int(size[0].strip())
-                        height = int(size[1].strip())
+                        width = float(size[0].strip())
+                        height = float(size[1].strip())
                         smaller_size = min(width, height)
-                        return int(smaller_size)
+                        return float(smaller_size)
         except (ValueError, AttributeError):
             print(Fore.RED + "||| Не смог получить меньший размер для сортировки |||" + Back.WHITE)
             return "!ERROR!"
